prevent flickering popups for elements that has positional styles
In the updatePopup, the popup position of x and y position is
wrapped between the html. If the elements have any style that changes
the position, padding, or margin, you'll get a flickering effect on the
mouse over.

This will fix that by setting the positions before rendering out the
html.