Added ability to specify an SVG filter for bubbles. To support this bubblesConfig now has a filterKey setting that will be applied to all bubbles. Each bubble can override this by setting their own filterKey property.
I've also added a new config section "filters" which works in the same way as the "fills" section but allows you to specify SVG filterss, though none are supplied out of the box.

To use a filter you must reference it by id and include it in an SVG element anywhere on the page.

Added an example page bubbles-with-shadows.html that shows how you can use this to add a drop shadow or glow like effect to the bubbles. Example also demonstrates existing ability to use gradient fills for bubbles. In the example I'm using radial gradients to apply a 3D like effect.